refactor: move database folder to infra
